+++ sync-for-commerce/src/codatsynccommerce/models/operations/update_visible_accounts_sync_flow.py
@@ -10,7 +10,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class UpdateVisibleAccountsSyncFlowRequest:
-    commerce_key: str = dataclasses.field(metadata={'path_param': { 'field_name': 'commerceKey', 'style': 'simple', 'explode': False }})
+    platform_key: str = dataclasses.field(metadata={'path_param': { 'field_name': 'platformKey', 'style': 'simple', 'explode': False }})
     visible_accounts: Optional[shared_visibleaccounts.VisibleAccounts] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})