+++ sync-for-commerce/RELEASES.md
@@ -519,3 +519,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.30.1] https://pypi.org/project/codat-sync-for-commerce/0.30.1 - synccommerce
+
+## 2023-09-01 16:29:32
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.77.1 (2.91.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.31.0] sync-for-commerce
+### Releases
+- [PyPI v0.31.0] https://pypi.org/project/codat-sync-for-commerce/0.31.0 - sync-for-commerce